First tests in language working
